requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
anthropic==0.7.8
pymupdf4llm==0.0.7
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3
python-pptx==0.6.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
